Develop server scaffolding
Still at a loss with the best way to do syntax highlighting...

Seems like our options are:
1. Writing a TextMate grammar, which has native support
2. Write a semantic tokenizer on the language server with tree-sitter
3. Same as the above but with the normal Bend parser.

2 and 3 will have slightly worse performance, but I think it's not that
bad; also, writing a new grammar in TextMate and mantaining it will be
too much trouble.

The normal Bend parser does not produce a concrete syntax tree, so
highlighting tokens like parentheses and commas won't be directly
possible. I think our best option right now is to do it with tree
sitter, but I haven't been able to find a project that does this directly.